Move platform-http-starter virtual threads tests to be integration tests
